@@ -61,7 +61,7 @@
         "file-saver": "^2.0.5",
         "history": "^4.9.0",
         "html-react-parser": "^2.0.0",
-        "i18next": "^17.0.11",
+        "i18next": "^21.9.1",
         "i18next-browser-languagedetector": "^3.0.3",
         "i18next-xhr-backend": "^3.2.2",
         "js-beautify": "^1.13.0",
@@ -70,12 +70,12 @@
         "mustache": "^4.2.0",
         "node-forge": "^0.10.0",
         "rc-tree": "^4.0.0-beta.2",
-        "react": "^16.9.0",
+        "react": "^18.2.0",
         "react-beautiful-dnd": "^12.2.0",
-        "react-dom": "^16.9.0",
+        "react-dom": "^18.2.0",
         "react-draggable": "^4.2.0",
         "react-helmet": "^5.2.1",
-        "react-i18next": "^10.11.5",
+        "react-i18next": "^11.18.5",
         "react-joyride": "^2.3.0",
         "react-notification-system": "^0.4.0",
         "react-redux": "^7.1.0",
@@ -105,9 +105,9 @@
         "@types/lodash-es": "^4.17.4",
         "@types/node": "^13.9.2",
         "@types/node-forge": "^0.9.3",
-        "@types/react": "^16.9.0",
+        "@types/react": "^18.0.18",
         "@types/react-beautiful-dnd": "^12.1.1",
-        "@types/react-dom": "^16.8.5",
+        "@types/react-dom": "^18.0.6",
         "@types/react-notification-system": "0.2.39",
         "@types/react-redux": "^7.1.1",
         "@types/react-router": "^5.1.18",
@@ -144,20 +144,7 @@
         "ts-node": "^10.8.1",
         "typescript": "^4.6.4"
     },
-    "peerDependencies": {
-        "react": "*"
-    },
     "browserslist": [
         "> 0.2%"
-    ],
-    "resolutions": {
-        "@types/react": "^16.9.0",
-        "@types/react-dom": "^16.8.5",
-        "@types/react-notification-system": "0.2.39"
-    },
-    "overrides": {
-        "@types/react": "^16.9.0",
-        "@types/react-dom": "^16.8.5",
-        "@types/react-notification-system": "0.2.39"
-    }
+    ]
 }
