@@ -49,7 +49,7 @@
         "@wso2is/validation": "^1.4.102",
         "axios": "^0.19.2",
         "history": "^4.9.0",
-        "i18next": "^17.0.11",
+        "i18next": "^21.9.1",
         "i18next-browser-languagedetector": "^3.0.3",
         "i18next-xhr-backend": "^3.2.2",
         "lodash-es": "^4.17.21",
@@ -57,10 +57,10 @@
         "path-browserify": "^1.0.1",
         "postcss-loader": "^3.0.0",
         "qrcode.react": "^1.0.0",
-        "react": "^16.9.0",
-        "react-dom": "^16.9.0",
+        "react": "^18.2.0",
+        "react-dom": "^18.2.0",
         "react-helmet": "^5.2.1",
-        "react-i18next": "^10.11.5",
+        "react-i18next": "^11.18.5",
         "react-notification-system": "^0.4.0",
         "react-redux": "^7.1.0",
         "react-router-dom": "^4.3.1",
@@ -77,7 +77,7 @@
         "@types/jest": "^26.0.14",
         "@types/lodash-es": "^4.17.4",
         "@types/node": "^13.9.2",
-        "@types/react-dom": "^16.8.5",
+        "@types/react-dom": "^18.0.6",
         "@types/react-notification-system": "0.2.39",
         "@types/react-redux": "^7.1.1",
         "@types/react-router": "^5.1.18",
@@ -113,15 +113,5 @@
     },
     "browserslist": [
         "> 0.2%"
-    ],
-    "resolutions": {
-        "@types/react": "^16.9.0",
-        "@types/react-dom": "^16.8.5",
-        "@types/react-notification-system": "0.2.39"
-    },
-    "overrides": {
-        "@types/react": "^16.9.0",
-        "@types/react-dom": "^16.8.5",
-        "@types/react-notification-system": "0.2.39"
-    }
+    ]
 }
