@@ -18,6 +18,7 @@
 
 import React, {
     ForwardRefExoticComponent,
+    PropsWithChildren,
     ReactElement,
     cloneElement,
     forwardRef,
@@ -45,161 +46,164 @@ export interface FormPropsInterface extends FormProps {
 
 /**
  * Implementation of the Form component.
- * @param props
+ * @param props - Props injected to the component.
  */
-export const Form: ForwardRefExoticComponent<FormPropsInterface> =
-    forwardRef((props: FormProps, ref): ReactElement => {
+export const Form: ForwardRefExoticComponent<PropsWithChildren<FormPropsInterface>> =
+    forwardRef((props: PropsWithChildren<FormProps>, ref): ReactElement => {
 
-    const { noValidate, triggerSubmit, ...other } = props;
-    const { children, onSubmit, uncontrolledForm, ...rest } = other;
+        const { noValidate, triggerSubmit, ...other } = props;
+        const { children, onSubmit, uncontrolledForm, ...rest } = other;
 
-    const formRef = useRef(null);
-    const childNodes = React.Children.toArray(children);
+        const formRef = useRef(null);
+        const childNodes = React.Children.toArray(children);
 
-    const skipFinalTypes = (type: String): boolean => {
+        const skipFinalTypes = (type: string): boolean => {
 
-        const typeToBeSkipped = [ "FieldDropdown" ]
+            const typeToBeSkipped = [ "FieldDropdown" ];
         
-        return typeToBeSkipped.some((skipType) => {
-            return type === skipType;
-        });
-    };
+            return typeToBeSkipped.some((skipType) => {
+                return type === skipType;
+            });
+        };
 
-    useImperativeHandle(ref, () => ({
-        triggerSubmit: () => {
-            /**
-             * What is this?
-             *
-             * Below this line we have the logic to trigger the native
-             * form's submit event. Currently we don't maintain UI for
-             * submit button, neither previous buttons.
-             *
-             * The parent wizard is only responsible for changing it's wizard
-             * pages based on the current index. To facilitate external
-             * submit handling we expose a imperative handler that uses
-             * a react ref to trigger the onSubmit manually. This is
-             * because, inherently react-final-form uses its form submission
-             * to properly keep track of the redux state.
-             *
-             * We should be able to synthetically tell react final form
-             * that a page has been changed and please trigger {@link handleSubmit}
-             * to make sure values are saved.
-             *
-             * {@see onSubmit}
-             * {@link https://final-form.org/docs/react-final-form/faq}
-             */
-            if (formRef) {
-                const submission = new Event('submit', {
-                    cancelable: true,
-                    bubbles: true
-                });
-                formRef.current?.dispatchEvent(submission);
+        useImperativeHandle(ref, () => ({
+            triggerSubmit: () => {
+                /**
+                 * What is this?
+                 *
+                 * Below this line we have the logic to trigger the native
+                 * form's submit event. Currently we don't maintain UI for
+                 * submit button, neither previous buttons.
+                 *
+                 * The parent wizard is only responsible for changing it's wizard
+                 * pages based on the current index. To facilitate external
+                 * submit handling we expose a imperative handler that uses
+                 * a react ref to trigger the onSubmit manually. This is
+                 * because, inherently react-final-form uses its form submission
+                 * to properly keep track of the redux state.
+                 *
+                 * We should be able to synthetically tell react final form
+                 * that a page has been changed and please trigger {@link handleSubmit}
+                 * to make sure values are saved.
+                 *
+                 * @see {@link onSubmit}
+                 * @see {@link https://final-form.org/docs/react-final-form/faq}
+                 */
+                if (formRef) {
+                    const submission = new Event("submit", {
+                        bubbles: true,
+                        cancelable: true
+                    });
+
+                    formRef.current?.dispatchEvent(submission);
+                }
             }
-        }
-    }));
+        }));
 
-    const addPropsToChild = (childNodes, formRenderProps: FormRenderProps) => {
+        const addPropsToChild = (childNodes, formRenderProps: FormRenderProps) => {
 
-        return childNodes.map((child: any) => {
+            return childNodes.map((child: any) => {
 
-            // Safety pre-conditions
-            if (!child) return null;
-            if (!child.type) return child;
+                // Safety pre-conditions
+                if (!child) return null;
+                if (!child.type) return child;
 
-            const { form, handleSubmit, pristine, submitting, values, initialValues } = formRenderProps;
+                const { form, handleSubmit, pristine, submitting, values, initialValues } = formRenderProps;
 
-            const parentFormProps = { form, handleSubmit, pristine, submitting, values, initialValues };
-            const childFieldProps = child.props;
-            const childProps: any = { childFieldProps, parentFormProps };
-            const allProps: any = { ...props, childProps }
+                const parentFormProps = { form, handleSubmit, pristine, submitting, values, initialValues };
+                const childFieldProps = child.props;
+                const childProps: any = { childFieldProps, parentFormProps };
 
-            // Check whether children of this element is valid
-            // and is type array.
-            const hasChildrenAndIsValid = Array.isArray(child.props?.children) &&
+                // Check whether children of this element is valid
+                // and is type array.
+                const hasChildrenAndIsValid = Array.isArray(child.props?.children) &&
                 child.props?.children.every(React.isValidElement) &&
                 child.props.children.length > 0;
 
-            // If the react element has only 1 child, the react top level
-            // API parses the children as a single object instead of type
-            // array.
-            const hasOnlyOneChild = (typeof child.props?.children === "object");
+                // If the react element has only 1 child, the react top level
+                // API parses the children as a single object instead of type
+                // array.
+                const hasOnlyOneChild = (typeof child.props?.children === "object");
 
-            if (uncontrolledForm) {
-                if ((hasChildrenAndIsValid || hasOnlyOneChild) && !skipFinalTypes(child.type.name)) {
-                    return React.createElement(child.type, {
-                        ...child.props,
-                        children: addPropsToChild(
-                            React.Children.toArray(child.props?.children),
-                            formRenderProps
-                        )
-                    });
+                if (uncontrolledForm) {
+                    if ((hasChildrenAndIsValid || hasOnlyOneChild) && !skipFinalTypes(child.type.name)) {
+                        return React.createElement(child.type, {
+                            ...child.props,
+                            children: addPropsToChild(
+                                React.Children.toArray(child.props?.children),
+                                formRenderProps
+                            )
+                        });
+                    }
                 }
-            }
-            return cloneElement(child, childProps);
 
-        });
+                return cloneElement(child, childProps);
 
-    }
+            });
 
-    const renderComponents = (childNodes, formRenderProps: FormRenderProps) => {
+        };
 
-        const modifiedChildNodes = addPropsToChild(childNodes, formRenderProps);
+        const renderComponents = (childNodes, formRenderProps: FormRenderProps) => {
 
-        return modifiedChildNodes.map((child: any, index: number) => {
-            if (!child) {
-                return null;
-            }
-            if (child.props.childFieldProps && child.props?.childFieldProps?.hidden) {
-                return null;
-            }
-            return (
-                <Grid.Row key={ index }>
-                    <Grid.Column width={ child.props.childFieldProps?.width }>
-                        { child }
-                    </Grid.Column>
-                </Grid.Row>
-            );
-        });
+            const modifiedChildNodes = addPropsToChild(childNodes, formRenderProps);
 
-    }
-
-    return (
-        <FinalForm
-            onSubmit={ (values, form) => onSubmit(values, form) }
-            keepDirtyOnReinitialize={ true }
-            render={ (formRenderProps: FormRenderProps) => {
-
-                const { handleSubmit } = formRenderProps;
-                if (triggerSubmit && typeof triggerSubmit === "function") {
-                    triggerSubmit(handleSubmit)
+            return modifiedChildNodes.map((child: any, index: number) => {
+                if (!child) {
+                    return null;
+                }
+                if (child.props.childFieldProps && child.props?.childFieldProps?.hidden) {
+                    return null;
                 }
 
                 return (
-                    <form
-                        noValidate={ noValidate }
-                        onSubmit={ handleSubmit }
-                        ref={ formRef }
-                    >
-                        <SemanticForm noValidate={ noValidate }>
-                            <Grid className="form-container with-max-width">
-                                { renderComponents(
-                                    childNodes,
-                                    formRenderProps
-                                ) }
-                            </Grid>
-                        </SemanticForm>
-                    </form>
+                    <Grid.Row key={ index }>
+                        <Grid.Column width={ child.props.childFieldProps?.width }>
+                            { child }
+                        </Grid.Column>
+                    </Grid.Row>
                 );
+            });
 
-            } }
-            { ...rest }
-        />
-    );
-});
+        };
+
+        return (
+            <FinalForm
+                onSubmit={ (values, form) => onSubmit(values, form) }
+                keepDirtyOnReinitialize={ true }
+                render={ (formRenderProps: FormRenderProps) => {
+
+                    const { handleSubmit } = formRenderProps;
+
+                    if (triggerSubmit && typeof triggerSubmit === "function") {
+                        triggerSubmit(handleSubmit);
+                    }
+
+                    return (
+                        <form
+                            noValidate={ noValidate }
+                            onSubmit={ handleSubmit }
+                            ref={ formRef }
+                        >
+                            <SemanticForm noValidate={ noValidate }>
+                                <Grid className="form-container with-max-width">
+                                    { renderComponents(
+                                        childNodes,
+                                        formRenderProps
+                                    ) }
+                                </Grid>
+                            </SemanticForm>
+                        </form>
+                    );
+
+                } }
+                { ...rest }
+            />
+        );
+    });
 
 /**
  * Default props for the component.
  */
 Form.defaultProps = {
-    noValidate: true,
+    noValidate: true
 };
